Move IR builders along with type definitions
