+++ src/IR/assignments.ts
@@ -1,4 +1,4 @@
-import { Expr, Identifier } from "./IR";
+import { Expr, id, Identifier } from "./IR";
 
 /**
  * Assignment statement of the form `variable = expr`. Raw OK
@@ -32,4 +32,37 @@ export interface OneToManyAssignment {
   type: "OneToManyAssignment";
   variables: Identifier[];
   expr: Expr;
+}
+
+export function assignment(
+  variable: Identifier | string,
+  expr: Expr
+): Assignment {
+  return {
+    type: "Assignment",
+    variable: typeof variable === "string" ? id(variable) : variable,
+    expr,
+  };
+}
+
+export function manyToManyAssignment(
+  variables: (Identifier | string)[],
+  exprs: Expr[]
+): ManyToManyAssignment {
+  return {
+    type: "ManyToManyAssignment",
+    variables: variables.map((v) => (typeof v === "string" ? id(v) : v)),
+    exprs,
+  };
+}
+
+export function oneToManyAssignment(
+  variables: (Identifier | string)[],
+  expr: Expr
+): OneToManyAssignment {
+  return {
+    type: "OneToManyAssignment",
+    variables: variables.map((v) => (typeof v === "string" ? id(v) : v)),
+    expr,
+  };
 }
